# portal/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Compliance Engine API
  version: 1.0.0
  description: |
    API for the Compliance Engine platform — ASV scanning, WAF management,
    Wazuh SIEM integration, and compliance assessment (SOC 2, PCI DSS, PCI SSS).

    ## Authentication

    All API requests require an API key passed in the `X-API-Key` header.
    Keys are scoped — see the `scopes` object on key creation for available
    permissions.

    ## Scopes

    | Scope | Grants |
    |-------|--------|
    | `read:scans` | Read scan results and status |
    | `write:scans` | Start and stop scans |
    | `read:waf` | Read WAF rules and traffic stats |
    | `manage:waf` | Create, update, delete WAF rules |
    | `read:siem` | Read SIEM alerts and agents |
    | `write:siem` | Resolve alerts, trigger active response |
    | `read:compliance` | Read compliance frameworks and assessments |
    | `admin` | Full access to all resources |

  contact:
    name: Compliance Engine Support
    email: support@compliance-engine.example
  license:
    name: Proprietary
servers:
  - url: https://api.compliance-engine.example/v1
    description: Production
  - url: https://sandbox.compliance-engine.example/v1
    description: Sandbox
security:
  - ApiKeyAuth: []
...
paths:
  /auth/api-keys:
    post:
      tags: [Auth]
      summary: Create API key
      operationId: createApiKey
      security:
        - ApiKeyAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Auth]
      summary: List API keys
      operationId: listApiKeys
      security:
        - ApiKeyAuth: [admin]
      responses:
        '200':
          description: List of API keys (secrets masked)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
  /auth/api-keys/{keyId}:
    parameters:
      - $ref: '#/components/parameters/keyId'
    get:
      tags: [Auth]
      summary: Get API key
      operationId: getApiKey
      security:
        - ApiKeyAuth: [admin]
      responses:
        '200':
          description: API key details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Auth]
      summary: Update API key
      operationId: updateApiKey
      security:
        - ApiKeyAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyUpdate'
      responses:
        '200':
          description: Updated API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Auth]
      summary: Revoke API key
      operationId: revokeApiKey
      security:
        - ApiKeyAuth: [admin]
      responses:
        '204':
          description: API key revoked
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/api-keys/{keyId}/rotate:
    parameters:
      - $ref: '#/components/parameters/keyId'
    post:
      tags: [Auth]
      summary: Rotate API key secret
      operationId: rotateApiKey
      security:
        - ApiKeyAuth: [admin]
      responses:
        '200':
          description: New secret returned (shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    keyId:
      name: keyId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Forbidden:
      description: Insufficient scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    Scope:
      type: string
      enum:
        - read:scans
        - write:scans
        - read:waf
        - manage:waf
        - read:siem
        - write:siem
        - read:compliance
        - admin
    ApiKeyCreate:
      type: object
      required: [name, scopes]
      properties:
        name:
          type: string
          example: My Integration Key
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
        expiresAt:
          type: string
          format: date-time
          description: Optional expiry. Omit for non-expiring.
    ApiKey:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        maskedKey:
          type: string
          example: sk_live_••••••••a1b2
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    ApiKeyCreated:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
          description: Full secret — shown only once.
          example: sk_live_EXAMPLE_PLACEHOLDER
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ApiKeyUpdate:
      type: object
      properties:
        name:
          type: string
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
...